python-sdk/pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "shelby-sdk"
version = "0.1.0"
description = "Python SDK for Shelby Protocol"
readme = "README.md"
requires-python = ">=3.11"
license = {text = "MIT"}

keywords = ["shelby", "sdk", "storage", "upload", "download", "blob", "account"]

authors = [
    {name = "Shelby Ecosystem", email = "dev@shelby.io"}
]

classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]

dependencies = [
    "httpx>=0.27.0",
    "pyyaml>=6.0",
    "aiofiles>=23.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-asyncio>=0.21.0",
    "black>=23.0",
    "mypy>=1.5",
    "ruff>=0.1.0",
]

[project.urls]
Homepage = "https://github.com/shelby-ecosystem/python-sdk"
Documentation = "https://shelby.io/docs/python-sdk"
Repository = "https://github.com/shelby-ecosystem/python-sdk"
Issues = "https://github.com/shelby-ecosystem/python-sdk/issues"

[tool.setuptools]
packages = ["shelby_sdk"]

[tool.black]
line-length = 100
target-version = ["py311"]

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
python_files = ["test_*.py"]
asyncio_default_fixture_loop_scope = "function"
